+++ packages/storybook/CHANGELOG.md
@@ -1,5 +1,12 @@
 # hawa-storybook
 
+## 0.26.71
+
+### Patch Changes
+
+- Updated dependencies
+  - @sikka/hawa@0.30.22
+
 ## 0.26.70
 
 ### Patch Changes